pyproject.toml
[tool.poetry]
name = "root_me_scoreboard_bot"
version = "1.0.0"
description = "Root-me scoreboard Discord bot"
authors = [
    "LighTender <87587438+LighTend3r@users.noreply.github.com>",
    "std3 <67806187+standard3@users.noreply.github.com>",
]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
poethepoet = "^0.29.0"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.38"
mkdocstrings = "^0.26.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poe.tasks]
test = "pytest"
start = "python3 root_me_scoreboard_bot/main.py"
docs = "mkdocs serve"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
